Move notifications in utils
